collection delete poup added

+++ components/delete-collection.tsx
@@ -1,17 +1,59 @@
-"use client"
-
+'use client'
+import * as React from 'react'
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle
+} from '@/components/ui/alert-dialog'
+import { toast } from 'sonner'
+import { IconSpinner } from '@/components/ui/icons'
 import { TrashIcon } from '@radix-ui/react-icons'
 import { deleteCollection } from '@/app/(chat)/collections/actions'
 
-export default function DeleteCollection({ id }: { id: string}) {
-    const handleClick = (id: string) => {
-        deleteCollection(id)
-      }
+export default function DeleteCollection({ id }: { id: string }) {
+  const [deleteDialogOpen, setDeleteDialogOpen] = React.useState(false)
+  const [isRemovePending, startRemoveTransition] = React.useTransition()
   return (
-    <div title="Delete" onClick={() => handleClick(id)}>
-    <TrashIcon
-      className="size-5 cursor-pointer"
-    />
-  </div>
+    <>
+      <div title="Delete" onClick={() => setDeleteDialogOpen(true)}>
+        <TrashIcon className="size-5 cursor-pointer" />
+      </div>
+      <AlertDialog open={deleteDialogOpen} onOpenChange={setDeleteDialogOpen}>
+        <AlertDialogContent>
+          <AlertDialogHeader>
+            <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
+            <AlertDialogDescription>
+              This will permanently delete your chat message and remove your
+              data from our servers.
+            </AlertDialogDescription>
+          </AlertDialogHeader>
+          <AlertDialogFooter>
+            <AlertDialogCancel disabled={isRemovePending}>
+              Cancel
+            </AlertDialogCancel>
+            <AlertDialogAction
+              disabled={isRemovePending}
+              onClick={event => {
+                event.preventDefault()
+                // @ts-ignore
+                startRemoveTransition(async () => {
+                  deleteCollection(id)
+                  setDeleteDialogOpen(false)
+                  toast.success('Collection deleted')
+                })
+              }}
+            >
+              {isRemovePending && <IconSpinner className="mr-2 animate-spin" />}
+              Delete
+            </AlertDialogAction>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialog>
+    </>
   )
 }
